requirements.txt
discord
pyyaml
py-dactyl
scikit-learn